+++ packages/back-end/src/models/FactTableModel.ts
@@ -10,10 +10,13 @@ import {
   UpdateColumnProps,
   UpdateFactTableProps,
 } from "../../types/fact-table";
+import { getConfigFactTables, usingFileConfig } from "../init/config";
+import { ALLOW_CREATE_FACT_TABLES } from "../util/secrets";
 
 const factTableSchema = new mongoose.Schema({
   id: String,
   organization: String,
+  official: Boolean,
   dateCreated: Date,
   dateUpdated: Date,
   name: String,
@@ -66,8 +69,17 @@ function toInterface(doc: FactTableDocument): FactTableInterface {
 }
 
 export async function getAllFactTablesForOrganization(organization: string) {
+  const factTables: FactTableInterface[] = [];
+
+  if (usingFileConfig()) {
+    const configFactTables = getConfigFactTables(organization);
+    factTables.push(...configFactTables);
+  }
+
   const docs = await FactTableModel.find({ organization });
-  return docs.map((doc) => toInterface(doc));
+  factTables.push(...docs.map((doc) => toInterface(doc)));
+
+  return factTables;
 }
 
 export type FactTableMap = Map<string, FactTableInterface>;
@@ -81,6 +93,14 @@ export async function getFactTableMap(
 }
 
 export async function getFactTable(organization: string, id: string) {
+  // First try looking in the config.yml file
+  if (usingFileConfig()) {
+    const configFactTables = getConfigFactTables(organization);
+    const configFactTable = configFactTables.find((f) => f.id === id);
+    if (configFactTable) return configFactTable;
+  }
+
+  // Fall back to looking in the database
   const doc = await FactTableModel.findOne({ organization, id });
   return doc ? toInterface(doc) : null;
 }
@@ -89,6 +109,10 @@ export async function createFactTable(
   organization: string,
   data: CreateFactTableProps
 ) {
+  if (usingFileConfig() && !ALLOW_CREATE_FACT_TABLES) {
+    throw new Error("Creating fact tables is not allowed");
+  }
+
   const doc = await FactTableModel.create({
     organization: organization,
     id: data.id || uniqid("ftb_"),
@@ -113,6 +137,10 @@ export async function updateFactTable(
   factTable: FactTableInterface,
   changes: UpdateFactTableProps
 ) {
+  if (factTable.official) {
+    throw new Error("Cannot update official fact tables");
+  }
+
   await FactTableModel.updateOne(
     {
       id: factTable.id,
@@ -132,6 +160,10 @@ export async function updateColumn(
   column: string,
   changes: UpdateColumnProps
 ) {
+  if (factTable.official) {
+    throw new Error("Cannot update official fact tables");
+  }
+
   const columnIndex = factTable.columns.findIndex((c) => c.column === column);
   if (columnIndex < 0) throw new Error("Could not find that column");
 
@@ -159,6 +191,10 @@ export async function createFactFilter(
   factTable: FactTableInterface,
   data: CreateFactFilterProps
 ) {
+  if (factTable.official) {
+    throw new Error("Cannot update official fact tables");
+  }
+
   const filter: FactFilterInterface = {
     id: data.id || uniqid("flt_"),
     name: data.name,
@@ -195,6 +231,10 @@ export async function updateFactFilter(
   filterId: string,
   changes: UpdateFactFilterProps
 ) {
+  if (factTable.official) {
+    throw new Error("Cannot update official fact tables");
+  }
+
   const filters = [...factTable.filters];
 
   const filterIndex = filters.findIndex((f) => f.id === filterId);
@@ -221,6 +261,10 @@ export async function updateFactFilter(
 }
 
 export async function deleteFactTable(factTable: FactTableInterface) {
+  if (factTable.official) {
+    throw new Error("Cannot delete official fact tables");
+  }
+
   await FactTableModel.deleteOne({
     id: factTable.id,
     organization: factTable.organization,
@@ -231,6 +275,10 @@ export async function deleteFactFilter(
   factTable: FactTableInterface,
   filterId: string
 ) {
+  if (factTable.official) {
+    throw new Error("Cannot update official fact tables");
+  }
+
   const newFilters = factTable.filters.filter((f) => f.id !== filterId);
 
   if (newFilters.length === factTable.filters.length) {
